Add file reader
